Refactored bot code, working on bot-db code

+++ BCSSBot/API/Models/UserUpdate.cs
@@ -5,7 +5,7 @@ namespace BCSSBot.API.Models
     public class UserUpdate
     {
         [JsonProperty("userHash")] 
-        public ulong UserHash { get; set; }
+        public int UserHash { get; set; }
 
         [JsonProperty("discordId")] 
         public ulong DiscordId { get; set; }
