+++ BCSSBot/Database/Models/Membership.cs
@@ -7,11 +7,11 @@ namespace BCSSBot.API.Models
     {
         [Key]
         [Column("user_hash")]
-        public ulong UserHash { get; set; }
+        public int UserHash { get; set; }
         public virtual User User { get; set; }
         [Key]
         [Column("permission_id")]
-        public ulong Id { get; set; }
+        public int Id { get; set; }
         public virtual Permission Permission { get; set; }
     }
 }
